# Avoid Django's lazy() when creating validators because it is too slow