@bluetech 2019-05-02
Here is the benchmark that I used: https://gist.github.com/bluetech/cdd3b3d6d9d8cfc7b1dfee3c6482679d