@tomchristie 2019-05-20
Okay, so this looks like some really good work here. It's also a *big* change footprint, which means that it's risky and difficult for us to take a call on.

Can you outline *exactly* what you think the behavioral changes would be as a result of this change?

The lazy translations provide the ability to deliver appropriately translated versions to the use on a per-request basis. Does this implementation change anything there that might not currently be caught by our test suite? When exactly are the translations made with this change? Can you walk me through exactly what's going on here?...

```python
self.validators.append(lambda value: MaxLengthValidator(
    self.max_length,
    message=self.error_messages['max_length'].format(max_length=self.max_length),
)(value))
```

(Initially confusing, because it looks like a lambda that's defined, but then immediatly called?)